requirements.txt
streamlit
python-dotenv
pandas
google-generativeai